pyproject.toml
[tool.poetry]
name = "edcontrol"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
tensorflow = "^2.13.0"
pypaz = "^0.2.5"
opencv-python = "^4.8.0.76"
numpy = "^1.25.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
